Download and parse custom data subset from the AirNow data API

This function uses the AirNow data webservice to retrieve subsets of data. This function is useful when users want data for a small subset of monitors within a particular region. It can also be used to access international data for countries outside the Canada-US-Mexico default region.

Datetimes can be anything that is understood by MazamaCoreUtils::parseDatetime(). For non-POSIXct values, the recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".

airnow_getDataCustom(
  parameterName = c("PM2.5"),
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  bbox = NULL,
  monitorType = c("both", "permanent", "mobile"),
  instrumentType = c("both", "monitor", "sensor"),
  baseUrl = "https://www.airnowapi.org/aq/data/"
)

Arguments

parameterName

One or more EPA AQS criteria parameter names.

starttime

Desired start datetime (ISO 8601).

endtime

Desired end datetime (ISO 8601).

timezone

Olson timezone used to interpret dates (required).

bbox

Bounding box vector defined as c(w, s, e, n).

monitorType

Specifies mobile or permanent monitor deployments.

instrumentType

Specifies regulatory monitors or low cost sensors.

baseUrl

Base URL for archived hourly data.

Value

Tibble of AirNow hourly data.

Examples

# \donttest{
library(AirMonitorIngest)

# Fail gracefully if any resources are not available
try({

  setAPIKey("airnow", "<YOUR_API_KEY>")

  # Guatemala City
  monitor <-
    airnow_getDataCustom(
      parameterName = "PM2.5",
      starttime = 20210101,
      endtime = 20220101,
      timezone = "UTC",
      bbox = c(-91.4, 13.7, -89.4, 15.7)
    )

}, silent = FALSE)
# }