Download and aggregate multiple hourly data files from AirNow

This function downloads, parses and combines hourly data files from baseUrl.

Parameters included in AirNow data include at least the following list:

BARPR

BC

CO

NO

NO2

NO2Y

NO2X

NOX

NOOY

OC

OZONE

PM10

PM2.5

PRECIP

RHUM

SO2

SRAD

TEMP

UV-AETH

WD

WS

Passing a vector of one ore more of the above names as the parameters argument will cause the resulting tibble to be filtered to contain only records for those parameters.

Datetimes can be anything that is understood by MazamaCoreUtils::parseDatetime(). For non-POSIXct values, the recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".

airnow_getHourlyData(
  parameterName = c("PM2.5"),
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  baseUrl = "https://files.airnowtech.org/airnow",
  verbose = FALSE
)

Arguments

parameterName
One or more EPA AQS criteria parameter names.

starttime
Desired start datetime (ISO 8601).

endtime
Desired end datetime (ISO 8601).

timezone
Olson timezone used to interpret dates (required).

baseUrl
Base URL for archived hourly data.

verbose
Logical flag controlling detailed progress statements.

Value

Tibble of AirNow hourly data.

Note

Hourly data are available from AirNow starting on 2014-01-01.

Data from locations whose timezones have a fractional offset from UTC are removed as the mts_monitor data model only supports data reported on hour boundaries. As of 2019-06-26, fractional offsets only applies to US Department of State monitors in Myanmar, Sri Lanka, India and Nepal.

Examples

if (FALSE) {
# Fail gracefully if any resources are not available
try({

library(AirMonitorIngest)

tbl <-
  airnow_getHourlyData(
    parameterName = "PM2.5",
    starttime = 2021101200,
    endtime = 2021101204,
    timezone = "America/Los_Angeles",
    verbose = TRUE
   )

}, silent = FALSE)
}