Download EPA air quality format description tables

Download tables with descriptions of data file formats. These are needed to properly interpret the data files downloaded from EPA. Tables are returned as a named list and include (on 2021-09-24):

"Site Description"

"Monitor Description"

"Annual Summary"

"Daily Summary"

"Hourly Data"

"8-Hour Average"

"Blanks Data"

For example, the `epa_aqs_parseHourlyData()` function uses the information in the table named "Hourly Data".

epa_aqs_getFormatDescriptionTables(
  url = "https://aqs.epa.gov/aqsweb/airdata/FileFormats.html"
)

Arguments

url
Character URL for EPA AQS file format descriptions.

Value

Named list of tables with format descriptions.

References

EPA AirData File Formats page

Examples

if (FALSE) {
library(AirMonitorIngest)

formatDescriptionTables <- epa_aqs_getFormatDescriptionTables()
head(formatDescriptionTables$`Hourly Data`)
}