Create 'meta' dnad 'data' dataframes from WRCC data

Create a meta dataframe with WRCC monitor metadata and a data dataframe with PM2.5 time series appropriate for use with the MazamaTimeSeries package.

The data model has monitor metadata stored in a tibble named meta. with a deviceDeploymentID unique identifier that is matched by column names in an associated data file.

wrcc_createMetaAndData(
  locationTbl = NULL,
  distanceThreshold = NULL,
  wrcc_data = NULL,
  unitID = NULL
)

Arguments

locationTbl
Tibble of "known locations" produced with MazamaLocationUtils.

distanceThreshold
Separation distance in meters between "known locations".

wrcc_data
Tibble of WRCC monitor data after QC and clustering have been applied.

unitID
WRCC station identifier (will be upcased).

Value

List with two tibbles.