Create a 'data' dataframe with required data

Create a local archive of EPA AQS transofrmed into data appropriate for use with the MazamaTimeSeries package.

The data model is that data are stored in a tibble named data with an initial datetime column and monitor-specific data in columns identified with a deviceDeploymentID that matches one found in the incoming meta table. Both data and meta will be saved as `.rda`, `.csv` and `.csv.gz` files for maximum portability to other languages like python, javascript or even to desktop spreadsheet software.

The archive structure underneath archiveBaseDir will be:

<archiveBaseDir>/EPA_AQS/<parameterCode>/<year>/

Files will be named as:

epa_aqs_<parameterCode>_<year>_meta.[rda|csv|csv.gz]
epa_aqs_<parameterCode>_<year>_data.[rda|csv|csv.gz]

epa_aqs_createLocalArchive(
  sites_locationTbl = NULL,
  downloadDir = NULL,
  parameterCode = NULL,
  years = NULL,
  archiveBaseDir = NULL,
  quiet = FALSE
)

Arguments

sites_locationTbl
EPA AQS "known locations" table for parameterCode monitors.

downloadDir
Directory in which to save downloaded EPA .zip files. If NULL, a temp directory is used.

parameterCode
EPA "Parameter Code".

years
Vector of years for which to transform data.

archiveBaseDir
Path under which the directory structure will be created.

quiet
Logical passed on to utils::download.file().

Value

Returns 0 on success.